+++ packages/backend/src/server/rest/hashtag/hashtags.ts
@@ -64,7 +64,7 @@ export const hashtagsListParamDef = z.object({
 export const hashtagsSearchParamDef = z.object({
 	limit: z.int().min(1).max(100).optional().default(10),
 	query: z.string(),
-	offset: z.int().optional().default(0),
+	offset: z.int().nonnegative().optional().default(0),
 });
 
 export const hashtagsShowParamDef = z.object({
@@ -251,7 +251,7 @@ export async function handleApiHashtagsShow(
 export const hashtagsUsersParamDef = z.object({
 	tag: z.string(),
 	limit: z.int().min(1).max(100).optional().default(10),
-	offset: z.int().optional().default(0),
+	offset: z.int().nonnegative().optional().default(0),
 	sort: z.enum(['+follower', '-follower', '+createdAt', '-createdAt', '+updatedAt', '-updatedAt']),
 	state: z.enum(['all', 'alive']).optional().default('all'),
 	origin: z.enum(['combined', 'local', 'remote']).optional().default('local'),